Add initial search page layout
